% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ART.R
\name{addWeightColumnNames}
\alias{addWeightColumnNames}
\title{Add column names to the weight matrix}
\usage{
addWeightColumnNames(network, columnNames)
}
\arguments{
\item{network}{The ART object}

\item{columnNames}{The column names to be added to the weight matrix.}
}
\value{
The weight matrix with the column names
}
\description{
Depending on the rule, add the column names in the data to the weight
matrix. For the fuzzy rule, the column names of the data and their complement (with _c)
attached to the names are added. For the hypersphere rule, the column names and the radius
column "R" are added. For ART1, the bottom up and top down weight columns are specified by
the tags "_bu" and "_td".
}
