% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ART.R
\name{ART}
\alias{ART}
\alias{isART}
\title{Create an ART Network}
\usage{
ART(
  rule = c("fuzzy", "hypersphere"),
  dimension,
  vigilance = 0.7,
  learningRate = 1,
  maxEpochs = 10
)

isART(net)
}
\arguments{
\item{rule}{The activation / match / learning rule to use. The choices available are fuzzy and hypersphere.}

\item{dimension}{The number of dimension/features in the data.}

\item{vigilance}{The vigilance parameter. Must be between 0 and 1.}

\item{learningRate}{The learning rate. Must be between 0 and 1.}

\item{net}{An object}
}
\value{
ART returns an ART object.

isART returns a logical value.
}
\description{
Create a new ART object

Check if the object is an ART
}
