% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ARTHypothesis.R
\name{ARTHypothesis}
\alias{ARTHypothesis}
\title{Linear Hypothesis Test for ART}
\usage{
ARTHypothesis(model, hypothesis.matrix, rhs = NULL)
}
\arguments{
\item{model}{an object of class \code{"artlm"}, usually, a result of a call to
\code{\link[=artlm]{artlm()}}.}

\item{hypothesis.matrix}{matrix (or vector) giving linear combinations of
coefficients by rows, or a character vector giving the hypothesis in
symbolic form (see \code{\link[car:linearHypothesis]{car::linearHypothesis()}})}

\item{rhs}{right-hand-side vector for hypothesis, with as many entries as
rows in the hypothesis matrix; can be omitted, in which case it defaults to
a vector of zeroes.}
}
\value{

}
\description{
This package allows for arbitrary hypothesis testing with ART. It requires
commands from the \code{car} library.
}
\seealso{
\code{\link[car:linearHypothesis]{car::linearHypothesis()}}
}
