% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ART.R
\name{ARTMAP}
\alias{ARTMAP}
\alias{isSimplified}
\alias{isARTMAP}
\title{Create an ARTMAP Network}
\usage{
ARTMAP(
  rule = c("fuzzy", "hypersphere", "art1"),
  dimension,
  vigilance = 0.7,
  learningRate = 1,
  maxEpochs = 10,
  simplified = TRUE,
  ...
)

isSimplified(net)

isARTMAP(net)
}
\arguments{
\item{rule}{The activation / match / learning rule to use. The choices available are fuzzy and hypersphere.}

\item{dimension}{The number of dimension/features in the data}

\item{vigilance}{The vigilance parameter. Must be between 0 and 1.}

\item{learningRate}{The learning rate. Must be between 0 and 1.}

\item{maxEpochs}{The maximum number of epochs. Default is 10.}

\item{simplified}{Logical. Whether to run the simplified version of ARTMAP. Default is FALSE.}

\item{...}{Other ART model specific parameters to be initialized.}

\item{net}{An object}
}
\value{
ARTMAP returns an ARTMAP object.

isSimplified returns a logical value.

isARTMAP returns a logical value.
}
\description{
Create a new ARTMAP object

Check if the ARTMAP object is the simplified version.

Check if the object is an ARTMAP
}
