% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ART.R
\name{colMax}
\alias{colMax}
\title{Find the maximum for each column}
\usage{
colMax(.data, na.rm = FALSE)
}
\arguments{
\item{.data}{A data object such as a data frame, a matrix, or an array.}

\item{na.rm}{Whether to remove NA from the calculation. Default is FALSE.}
}
\value{
A vector of maximum values for each column
}
\description{
Find the maximum for each column
}
