% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ART.R
\name{colMin}
\alias{colMin}
\title{Find the minimum for each column}
\usage{
colMin(.data, na.rm = FALSE)
}
\arguments{
\item{.data}{A data object such as a data frame, a matrix, or an array.}

\item{na.rm}{Whether to remove NA from the calculation. Default is FALSE.}
}
\value{
A vector of minimum values for each column
}
\description{
Find the minimum for each column
}
