% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/artlm.R
\name{confint.artlm}
\alias{confint.artlm}
\title{Confidence Intervals for Linear ART Parameters}
\usage{
\method{confint}{artlm}(object, parm, nrg = 10000, level = 0.95)
}
\arguments{
\item{parm}{a specification of which parameters are to be given confidence
intervals, either a vector of numbers or a vector of names. If missing,
all parameters are considered.}

\item{nrg}{an \code{integer} that represents the number of random
permutations to calculate when there are more than 10 groups. Defaults
to 10,000.}

\item{level}{the confidence level required. Defaults to 95\\%}

\item{an}{object of class \code{"artlm"}, usually, a result of a call to
\code{\link{artlm}}.}
}
\value{
A matrix (or vector) with columns giving lower and upper confidence
limits for each parameter. These will be labelled as (1-level)/2 and
1 - (1-level)/2 in \\% (by default 2.5\\% and 97.5\\%).
}
\description{
Confidence Intervals for Linear ART Parameters
}
