% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ART.R
\name{createDummyCodeMap}
\alias{createDummyCodeMap}
\alias{encodeLabel}
\alias{decode}
\title{Make Dummy Code}
\usage{
createDummyCodeMap(classLabels)

encodeLabel(classLabels, dummyCode)

decode(dummyClasses, dummyCode)
}
\arguments{
\item{classLabels}{A vector of class labels. They must be numeric.}

\item{dummyCode}{The mapping that converts the class labels to dummy code and vice versa.}

\item{dummyClasses}{The matrix of the dummy codes associated with the data points}
}
\value{
A list of dummy codes with the keys being the class labels and the values being the dummy (binary) codes.

A matrix of binary values

A vector of class labels
}
\description{
Generate the dummy codes for all the possible class labels

Convert numeric code to binary dummy code

Convert dummy code back to the class label
}
