% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backend.R
\name{CRS.bin}
\alias{CRS.bin}
\title{Hypothesis Test for ART}
\usage{
CRS.bin(lambda, c.beta, G, alpha = 0.05, nj = 1)
}
\arguments{
\item{lambda}{scalar for the null hypothesis c'beta = lambda (default 0)}

\item{c.beta}{vector of parameters entering the null hypothesis c'beta - with one estimator per cluster (q x 1)}

\item{G}{the group of all trasformations (use random.G to get)}

\item{alpha}{significance level (dafault 0.05)}

\item{nj}{q x 1 vector of sample sizes in each cluster (for alternative weighting)}
}
\value{
A \code{logical} that is true if the test rejects the null hypothesis.
}
\description{
This helper function only determines whether the test passes. It is used by \code{\link[=CRS.CI]{CRS.CI()}}.
}
