% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backend.R
\name{CRS.CI}
\alias{CRS.CI}
\title{Confidence Intervals for ART}
\usage{
CRS.CI(c.beta, G, alpha = 0.05, nj = 1)
}
\arguments{
\item{c.beta}{vector of parameters entering the null hypothesis c'beta
- with one estimator per cluster (q x 1)}

\item{G}{the group of all trasformations (use random.G to get)}

\item{alpha}{significance level (dafault 0.05)}

\item{nj}{q x 1 vector of sample sizes in each cluster (for alternative weighting)}
}
\value{
The 1-alpha confidence interval for c'beta
}
\description{
This helper function computes confidence intervals by test inversion using bisection.
}
