% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{drawWeight.fuzzy}
\alias{drawWeight.fuzzy}
\title{drawWeight.fuzzy}
\usage{
\method{drawWeight}{fuzzy}(w, classLabels = NULL)
}
\arguments{
\item{w}{The weight matrix}

\item{classLabels}{The vector of class labels associated with all weights in F2}
}
\value{
A ggplot
}
\description{
Draw the 2D rectangular weights created by the fuzzy rule
}
