% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{drawWeight.hypersphere}
\alias{drawWeight.hypersphere}
\title{drawWeight.hypersphere}
\usage{
\method{drawWeight}{hypersphere}(w, classLabels = NULL)
}
\arguments{
\item{w}{The data frame with the first 2 columns being x and y coordinates of the weight, and the third column being the radii.}

\item{classLabels}{The vector of class labels associated with the weights in F2.}
}
\description{
Plot 2D circles
}
