% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{drawWeight}
\alias{drawWeight}
\title{drawWeight}
\usage{
drawWeight(x, ...)
}
\arguments{
\item{x}{An ART object}

\item{...}{Additional arguments to be passed}
}
\description{
Generic function for plotting weights
}
