% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ART.R
\name{getModuleById}
\alias{getModuleById}
\title{Select a Module}
\usage{
getModuleById(net, id)
}
\arguments{
\item{net}{An ART network}

\item{id}{The id of the module. Either a numeric or a character/string id.}
}
\value{
An ART module
}
\description{
Get the module from the network by the id
}
