% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ART.R
\name{getTopoClustersCategories}
\alias{getTopoClustersCategories}
\title{Topological ART Clusters}
\usage{
getTopoClustersCategories(module)
}
\arguments{
\item{module}{A TopoART module}
}
\value{
A vector of linked cluster categories
}
\description{
Return all linked cluster categories created by a TopoART module
}
