% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mqy2015}
\alias{mqy2015}
\title{Data from Meng, Qian and Yared (2015).}
\format{
A data frame with 689 rows and 8 variables:
\describe{
\item{province}{name of province}
\item{year}{the year}
\item{lgrain}{offical government issued }
\item{lgrain_pred}{predicted log grain production (to correct misreports)}
\item{ldeath_b}{log number of deaths in the following year}
\item{lurbpop}{log urban population}
\item{ltotpop}{log total population}
\item{main}{logical for provinces included in main sample of 19 provinces}
}
}
\source{
\url{https://doi.org/10.1093/restud/rdv016}
}
\usage{
mqy2015
}
\description{
A panel dataset of 23 Chinese provinces from 1953-1982 with data on official
and constructed grain production. This data is used to show that food
production and mortality were correlated during the great famine which occurs
from 1958-1960.
}
\keyword{datasets}
