% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mr2015}
\alias{mr2015}
\title{Data from Munyo and Rossi (2015).}
\format{
A data frame with 2631 rows and 12 variables:
\describe{
\item{index}{a time running variable and unique index entry}
\item{totalcrime}{daily number of reported felonies}
\item{releases}{daily releases from ComCar, the main prison in Montevideo}
\item{temperature}{average temparature in Celsius}
\item{rainfall}{rainfall in millimeters}
\item{holiday}{indicator variable for Uruguay national holiday}
\item{sunshine}{hours of sunshine}
\item{endofyear}{an indicator for the last day of the year (New Years Eve)}
\item{dayofweek}{the day of the week}
\item{year}{the year}
\item{month}{integer representation of the month}
\item{date}{day of the month}
}
}
\source{
\url{https://doi.org/10.1016/j.jpubeco.2014.12.002}
}
\usage{
mr2015
}
\description{
A dataset that includes the universe of criminal incidents reported at the
Police Department of Montevideo: more than 690,000 felonies reported in
Montevideo, the capital of Uruguay, between January 1st 2004 and
March 15th 2011 (2631 days). This data is for academic use only.
}
\keyword{datasets}
