% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ART.R
\name{normalize}
\alias{normalize}
\title{Data Normalization}
\usage{
normalize(.data)
}
\arguments{
\item{.data}{A data object such as a data frame, a matrix , or an array.}
}
\value{
A normalized data object
}
\description{
Normalize each column of data to values between 0 and 1
}
