% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot.ARTMAP}
\alias{plot.ARTMAP}
\title{plot.ARTMAP}
\usage{
\method{plot}{ARTMAP}(net, .data, classes = NULL, dummyCodeMap = NULL)
}
\arguments{
\item{net}{An ARTMAP network}

\item{.data}{The data frame used for plotting the data. It must have 2 columns specifying x and y values.}

\item{classes}{The class labels associated with the data points.}

\item{dummyCodeMap}{The dummy code rule that maps each class label to its dummy code.}
}
\value{
A ggplot
}
\description{
Plot a 2 dimensional Fuzzy ARTMAP category distribution
}
