% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ART.R
\name{predict.ART}
\alias{predict.ART}
\title{ART Prediction}
\usage{
\method{predict}{ART}(network, id, .data)
}
\arguments{
\item{network}{An ART object}

\item{id}{The id of the module}

\item{.data}{The data used for prediction/testing. The data must be normalized between 0 and 1.}
}
\value{
Returns a list containing the predicted F2 categories.
}
\description{
The ART prediction/classification method
}
