% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ART.R
\name{predict.ARTMAP}
\alias{predict.ARTMAP}
\title{ARTMAP Prediction}
\usage{
\method{predict}{ARTMAP}(network, .data, classLabels = NULL, dummyClasses = NULL)
}
\arguments{
\item{network}{An ARTMAP object}

\item{.data}{The data used for training. The data must be normalized between 0 and 1.}

\item{classLabels}{The classes for each data point. They mustx be numeric values.}

\item{dummyClasses}{The classes for each data point. These are dummy codes for the labels and must be binary.}
}
\value{
Returns a list containing three items: 1. categories - the mapfield categories predicted, 2. category_a - the F2 categories predicted, and 3. matched - whether the mapfield categories predicted match the actual values.
}
\description{
The ARTMAP prediction/classification method
}
