% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ART.R
\name{predict.TopoART}
\alias{predict.TopoART}
\title{Topological ART Prediction}
\usage{
\method{predict}{TopoART}(network, id, .data)
}
\arguments{
\item{network}{A TopoART object}

\item{id}{The id of the module}

\item{.data}{The data used for prediction. The data must be normalized between 0 and 1.}
}
\value{
Returns a list containing the predicted F2 categories and the linked clusters.
}
\description{
The TopoART prediction/classification method
}
