% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rart.R
\name{random.G}
\alias{random.G}
\title{Groups of Transformations for ART}
\usage{
random.G(q, B = 10000)
}
\arguments{
\item{q}{the dimension of the data / # of clusters (q point estimators)}

\item{B}{the number of random draws if q>10}
}
\value{
A B x q matrix of transformations (G)
}
\description{
This helper function computes the group G of transformations for the CRS test.
}
