% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ART.R
\name{train.ART}
\alias{train.ART}
\title{Train an ART network}
\usage{
\method{train}{ART}(network, .data)
}
\arguments{
\item{network}{An ART  object}

\item{.data}{The data used for training.}
}
\description{
The ART training method
}
