% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ART.R
\name{train.TopoART}
\alias{train.TopoART}
\title{Train a Topological ART Network}
\usage{
\method{train}{TopoART}(network, .data)
}
\arguments{
\item{network}{An TopoART  object}

\item{.data}{The data used for training.}
}
\description{
The TopoART training method
}
