% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/artlm.R
\name{vcov.artlm}
\alias{vcov.artlm}
\title{Dummy vcov for ART}
\usage{
\method{vcov}{artlm}(...)
}
\arguments{
\item{...}{variables that are passed to \code{\link[=vcov.lm]{vcov.lm()}}}
}
\value{
A matrix of \code{NA} values in the same shape as a variance
covariance matrix.
}
\description{
ART does not generate standard errors or a variance covariance matrix. So,
this function will always return \code{NA}.
}
