% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ART.R
\name{ART}
\alias{ART}
\alias{isART}
\title{ART}
\usage{
ART(
  numModules = 1,
  rule = c("fuzzy", "hypersphere", "ART1"),
  dimension,
  vigilance = 0.7,
  learningRate = 1,
  maxEpochs = 10,
  ...
)

isART(net)
}
\arguments{
\item{numModules}{The number of modules in the network. Currently, it supports only 1 module.}

\item{rule}{The activation / match / lerning rule to use. The choices available are fuzzy.}

\item{dimension}{The number of dimension/features in the data.}

\item{vigilance}{The vigilance parameter. Must be between 0 and 1.}

\item{learningRate}{The learning rate. Must be between 0 and 1.}

\item{maxEpochs}{The maximum number of epochs to run. Default is 10.}

\item{...}{Other ART model specific parameters to be initialized.}

\item{net}{An object}
}
\value{
isART returns a logical value.
}
\description{
Create a new ART object

Check if the object is an ART
}
