% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linearHypothesis.R
\name{ARTHypothesis}
\alias{ARTHypothesis}
\title{Title}
\usage{
ARTHypothesis(model, hypothesis.matrix, rhs = NULL, coef. = coef(model), ...)
}
\arguments{
\item{...}{}
}
\value{

}
\description{
Title
}
\seealso{
\code{\link[car:linearHypothesis]{car::linearHypothesis()}}
}
