% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{drawWeight}
\alias{drawWeight}
\title{drawWeight}
\usage{
drawWeight(w, x, y, ...)
}
\arguments{
\item{w}{The weight object}

\item{x}{The x axis variable}

\item{y}{The y axis variable}

\item{...}{Additional arguments to be passed}
}
\description{
Generic function for plotting weights
}
