% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot.ART}
\alias{plot.ART}
\title{plot.ART}
\usage{
\method{plot}{ART}(net, id = 0, .data)
}
\arguments{
\item{net}{An ART network}

\item{id}{The numeric id of the module to be plotted. The base module ART a (default) has an id = 0. The ids of the higher order module are incremented.}

\item{.data}{The data frame used for plotting the data. It must have 2 columns corresponding to the x and y values, respectively.}
}
\value{
A ggplot
}
\description{
Plot a 2 dimensional Fuzzy ART category distribution
}
