% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ART.R
\name{TopoART}
\alias{TopoART}
\title{Create a Topological ART Network}
\usage{
TopoART(
  rule = c("fuzzy", "hypersphere"),
  dimension,
  numModules = 2,
  vigilance = 0.7,
  learningRate1 = 1,
  learningRate2 = 0.6,
  tau = 100,
  phi = 6,
  maxEpochs = 20
)
}
\arguments{
\item{rule}{The activation / match / lerning rule to use. The choices available are fuzzy and hypersphere.}

\item{dimension}{The number of dimension/features in the data}

\item{numModules}{The number of TopoART modules. The default is 2.}

\item{vigilance}{The vigilance parameter. Must be between 0 and 1.}

\item{learningRate1}{The learning rate when this neuron has the highest activation. Must be between 0 and 1.}

\item{learningRate2}{The learning rate when this neuron has the second highest activation. Must be between 0 and 1.}

\item{tau}{The number of learning cycles before the F2 neurons with counts below the threshold are removed. Default is 100.}

\item{phi}{The count threshold. The F2 neuron is kept when the count >= phi, removed when the count < phi.}

\item{maxEpochs}{The maximum number of epochs. Default is 20.}
}
\value{
TopoART returns a TopoART object.
}
\description{
Create a new TopoART object
}
