% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ART.R
\name{train.ARTMAP}
\alias{train.ARTMAP}
\title{Train an ARTMAP Network}
\usage{
\method{train}{ARTMAP}(network, .data, target)
}
\arguments{
\item{network}{An ARTMAP object}

\item{.data}{The data used for training.}

\item{target}{Either a numeric vector or a matrix. Use the vector form when running the simplified ARTMAP classification. Use the matrix 
form when running the standard ARTMAP classification where the target labels must be binary values. For regression which requires the 
standard ARTMAP, either a vector or a matrix (single column) of continuous values (normalized between 0 and 1) can be used.}
}
\value{
The ARTMAP object
}
\description{
The ARTMAP training method
}
