% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ART.R
\name{train.ARTMAP}
\alias{train.ARTMAP}
\title{Train an ARTMAP Network}
\usage{
\method{train}{ARTMAP}(network, .data, classLabels = NULL, dummyClasses = NULL)
}
\arguments{
\item{network}{An ARTMAP object}

\item{.data}{The data used for training.}

\item{classLabels}{The classes for each data point. They are numeric values.}

\item{dummyClasses}{The classes for each data point. These are dummy codes for the labels and must be binary.}

\item{simplified}{Logical. Whether to use the simplified ART network. Default is FALSE.}
}
\description{
The ARTMAP training method
}
